Time: 0 ms (100.00%) | Memory: 21.9 MB (75.43%) - LeetSync

+++ 98-validate-binary-search-tree/validate-binary-search-tree.cpp
@@ -19,7 +19,7 @@ public:
 
         while(cur){
             if(cur->left==NULL){
-                if((long long)prev_val>=(long long)cur->val)ans=false;
+                if(prev_val>=cur->val)ans=false;
                 prev_val=cur->val;
 
                 cur=cur->right;
